fix nonce type

+++ contracts/periphery/contracts/s2s/SmartChainXApp.sol
@@ -106,5 +106,5 @@ abstract contract SmartChainXApp {
     /// @param lane Lane id
     /// @param nonce Nonce of the callback message
     /// @param result Dispatch result of cross chain message
-    function onMessageDelivered(bytes4 lane, uint256 nonce, bool result) external virtual;
+    function onMessageDelivered(bytes4 lane, uint64 nonce, bool result) external virtual;
 }
